Add support for ADS1115
Since the ADS1015 and ADS1115 are *very* similar this change adds support for the latter.

The ADS1115 uses a different set of sample rates to the ADS1015 and supports 16-bit data output.

An auto-detect function is supplied that uses the average timing of a set of ADC reads in order to detect which device is connected and adjust the library accordingly.